+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## dev
 
 - Fix `SystemTracker` and `PidTracker` to not print dummy stats on start when header is disabled
+- Add optional `start_time` parameter to `SystemTracker` and `PidTracker`
 ## v0.3.1 (May 30, 2025)
 
 - Generate card for failed step